pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gemini-gradio"
version = "0.0.2"
description = "A Python package for creating Gradio applications with Google Gemini models"
authors = [
    { name = "AK", email = "ahsen.khaliq@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "gradio>=5.5.0",
    "google-generativeai",
]

[project.urls]
homepage = "https://github.com/AK391/gemini-gradio"
repository = "https://github.com/AK391/gemini-gradio"

[project.optional-dependencies]
dev = ["pytest"]

[tool.hatch.build.targets.wheel]
packages = ["gemini_gradio"]

